requirements.txt
langchain
langchain-community
langchain-core
pypdf
tiktoken
python-multipart
PyPDF2
faiss-cpu
langchain_groq
unstructured
python-docx
python-dotenv
streamlit